% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots accidents locations on a map for a given year and a given state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value used to specify a state number.}

\item{year}{A numeric value to specify the year for which the graphic
is plotted. An error will be thrown if there are no records for the given
state and the specified year. Wrong year or state number specification will
produce an error.}
}
\value{
A map plot of accidents location is returned.
}
\description{
Plots accidents locations on a map for a given year and a given state.
}
\examples{
\dontrun{
 fars_map_state(1,2013)
 }

}
