% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Month and year columns from data files of given years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric or character vector containing years for which, the data
will be read.}
}
\value{
A list of tbl data frames with columns "MONTH" and "year" corresponding
to the specified years. In case of wrong year specification an error will be 
thrown.
}
\description{
This function reads the data from files for a set of given years. For each year
it reads the data, adds a "year" column and selects the two columns "MONTH" and
"year".
}
\examples{
fars_read_years(2013)
fars_read_years(2013:2014)

}
