% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data by year and by month for a given set of years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric or character vector containing years for which, the data
will be read.}
}
\value{
A tbl data frame containing the number of accidents summarized by year 
and by month.Bad year argument or missing files will produce an error.
}
\description{
This function reads years and months from data for a given set of years.
The number of accidents by year and by month.
}
\examples{
fars_summarize_years("2013")
fars_summarize_years(2013:2015)

}
