% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make file names corresponding to given years.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or a character vector that specifies the years 
for which we want to build the file names.}
}
\value{
A character vector containing the file names
}
\description{
This function creates the file names corresponding to a given vector of years.
}
\examples{
make_filename(2013)
make_filename("2013")
make_filename(2013:2015)
}
